% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_rgb.R
\name{col_to_int}
\alias{col_to_int}
\title{Fast conversion of colors-as-strings to an integer vector of packed RGBA values.}
\usage{
col_to_int(col)
}
\arguments{
\item{col}{Character vector of color names. Supports all R color names
(e.g. "red", "hotpink") and hex colors of the form:  "#RRGGBBAA", 
"#RRGGBB", "#RGBA" and "#RGB".}
}
\value{
Integer vector where the 4 bytes making up each integer represent 
  the RGBA values of the colour
}
\description{
Packed RGBA values are used in R's native raster objects.
}
\examples{
col_to_int(c("hotpink", "#abc", "#aabb99", "#aabb9980"))
}
