% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_rgb.R
\name{col_to_packed_int}
\alias{col_to_packed_int}
\title{Convert colours-as-strings to packed integer colours for nativeRasters}
\usage{
col_to_packed_int(col)
}
\arguments{
\item{col}{Character vector of colour names. Supports all R colour names
(e.g. "red", "hotpink") and hex colours of the form:  "#RRGGBBAA", 
"#RRGGBB", "#RGBA" and "#RGB".}
}
\value{
integer vector
}
\description{
Convert colours-as-strings to packed integer colours for nativeRasters
}
