% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_rgb.R
\name{col_to_rgb}
\alias{col_to_rgb}
\title{Convert colours-as-strings to a matrix of RGBA integers}
\usage{
col_to_rgb(col)
}
\arguments{
\item{col}{Character vector of colour names. Supports all R colour names
(e.g. "red", "hotpink") and hex colours of the form:  "#RRGGBBAA", 
"#RRGGBB", "#RGBA" and "#RGB".}
}
\description{
This is a proof-of-concept drop-in replacement for \code{grDevices::col2rgb()}
which uses an order-preserving minimal perfect hash to lookup R colour
names.
}
\examples{
col_to_rgb(c("hotpink", "#abc", "#aabb99", "#aabb9980"))


}
